all reimplemented, stable
